https://github.com/hasii2011/PyUt/issues/6

+++ packageme.sh
@@ -1,9 +1,20 @@
 #!/usr/bin/env bash
-
+#
+#  Assumes python 3 is on PATH
+#
 clear
-echo "Clean up old stuff"
-rm -rf build dist
 
-python3 setup.py py2app --iconfile src/org/pyut/resources/img/Pyut.icns
-
+if [[ $# -eq 0 ]] ; then
+        echo "in alias mode"
+        rm -rf build dist
+        python3 setup.py py2app -A --iconfile src/org/pyut/resources/img/Pyut.icns
+else
+    if [[ ${1} = 'deploy' ]] ; then
+            echo "in deploy mode"
+            rm -rf build dist
+            python3 setup.py py2app --iconfile src/org/pyut/resources/img/Pyut.icns
+    else
+        echo "Unknown command line arguments"
+    fi
 rm -rf src/UNKNOWN.egg-info
+fi
